Directories

+++ src/main/java/ru/spsuace/homework4/files/Directories.java
@@ -17,37 +17,31 @@ public class Directories {
     public static int removeWithFile(String path) {
 
         int count = 0;
-        File dir1 = new File(path);
-            if (!dir1.exists()) {
+        File main = new File(path);
+        if (!main.exists() && !main.isDirectory()) {
             return count;
-            }
-        File[] listFiles = dir1.listFiles();
-        for (int i = 0; i < listFiles.length; i++) {
-            if (listFiles[i].isFile()) {
-                listFiles[i].delete();
-                count += 1;
-            }
-            if (listFiles[i].isDirectory()) {
-                count += removeWithFile(listFiles[i].getPath());
-            }
         }
-        dir1.delete();
-        count += 1;
+        File[] listFiles = main.listFiles();
+        if (listFiles == null) {
+            count = 1;
+        } else {
+            for (int i = 0; i < listFiles.length; i++) {
+                if (listFiles[i].isFile()) {
+                    listFiles[i].delete();
+                    count += 1;
+                }
+                if (listFiles[i].isDirectory()) {
+                    count += removeWithFile(listFiles[i].getPath());
+                }
+            }
+            main.delete();
+            count += 1;
+        };
         return count;
     }
 }
     /**
      * С использованием Path
      */
-   // public static int removeWithPath(String path) {
-// Path start = Paths.get(path);
-// DeleterVisitor deletePaths = new DeleterVisitor();
-// try {
-// Files.walkFileTree(start, deletePaths);
-// } catch (IOException exception) {
-// }
-//
-// return deletePaths.countFilesAndDirectories;
-// }
